Use undici for fetch for node 16 support 😒
Node 16.17.0 suppots fetch, but only with the `--experimental-fetch`
flag and there's no easy way to pass that in to stuff when running npm
scripts or npm bin files, and since Node's fetch implementation is
powered by undici and we already need it for TypeScript typings, let's
just pull in undici as a dependency.

This annoyingly has the side effect that it will probably make stubbing
fetch much easier in unit tests, since we'll be stubbing an import
rather than a global.